BAU: Refactor credential issue endpoint artefacts
Move classes needed for the issue endpoint from lib into
the veriable credential module in other to make lib more self contained
