+++ lambdas/issuecredential/src/main/java/uk/gov/di/ipv/cri/address/api/exception/CredentialRequestException.java
@@ -1,14 +1,10 @@
-package uk.gov.di.ipv.cri.address.library.exception;
+package uk.gov.di.ipv.cri.address.api.exception;
 
 import uk.gov.di.ipv.cri.address.library.error.ErrorResponse;
 
 public class CredentialRequestException extends Exception {
     public CredentialRequestException(ErrorResponse invalidRequestParam) {
         super(invalidRequestParam.getMessage());
-    }
-
-    public CredentialRequestException(String message) {
-        super(message);
     }
 
     public CredentialRequestException(String message, Exception e) {
