+++ lambdas/issuecredential/src/test/java/uk/gov/di/ipv/cri/address/api/helpers/KMSSignerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.cri.address.library.helpers;
+package uk.gov.di.ipv.cri.address.api.helpers;
 
 import com.amazonaws.services.kms.AWSKMS;
 import com.amazonaws.services.kms.model.MessageType;
@@ -17,6 +17,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.di.ipv.cri.address.library.helpers.KMSSigner;
 
 import java.nio.ByteBuffer;
 import java.util.Map;
