+++ lambdas/issuecredential/src/test/java/uk/gov/di/ipv/cri/address/api/helpers/SignClaimSetJwtTest.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.cri.address.library.helpers;
+package uk.gov.di.ipv.cri.address.api.helpers;
 
 import com.nimbusds.jose.JOSEException;
 import com.nimbusds.jose.crypto.ECDSASigner;
@@ -9,7 +9,8 @@ import com.nimbusds.jwt.SignedJWT;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.di.ipv.cri.address.library.helpers.fixtures.TestFixtures;
+import uk.gov.di.ipv.cri.address.api.helpers.fixtures.TestFixtures;
+import uk.gov.di.ipv.cri.address.library.helpers.SignClaimSetJwt;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.spec.InvalidKeySpecException;
